+++ Engine/Source/Programs/UnrealCloudDDC/changelog.md
@@ -1,4 +1,5 @@
 # Unreleased
+* Improved handling of cassandra timeouts during ref GC.
 
 # 0.1.0
 * First public release of UnrealCloudDDC